some work on the 'latest reviews' and 'opening this week' use case
Former-commit-id: 40850f6be6b8a89b4453572988a10463c823a75b [formerly 1e2fe8c6d471b9a577e7caea4f300b8b23971908]
Former-commit-id: 0e9cdca27dfb38acf67fdf6540fbcdd849a9de27

+++ Domain_Objects/src/main/java/potatoes/project/controller/UserController.java
@@ -80,7 +80,7 @@ public class UserController {
         
     @ResponseBody
     @GetMapping("/users/{id}")
-    public ModelAndView getUser(@PathVariable long id, Model model){
+    public ModelAndView getUser(@PathVariable int id, Model model){
     	ModelAndView mav = new ModelAndView();
     	User toGet = userService.findByUserID(id);
     	mav.setViewName("profile");
