+++ Domain_Objects/src/main/java/potatoes/project/service/UserService.java
@@ -33,7 +33,7 @@ public class UserService {
 	public User findByEmail(String email) {
 		return userRepository.findByEmail(email);
 	}
-	public User findByUserID(long id) {
+	public User findByUserID(int id) {
 		return userRepository.findByUserID(id);
 	}
 }
